+++ xmake/modules/core/tools/nvcc.lua
@@ -27,7 +27,7 @@ import("core.platform.platform")
 import("core.language.language")
 import("core.project.policy")
 import("utils.progress")
-import("private.tools.ccache")
+import("private.tools.compiler_cache")
 
 -- init it
 function init(self)
@@ -286,12 +286,13 @@ function compargv(self, sourcefile, objectfile, flags)
     return self:program(), table.join("-c", flags, "-o", objectfile, sourcefile)
 end
 
-function ccache_compargv(self, sourcefile, objectfile, flags)
-    return ccache.cmdargv(self:program(), table.join("-c", flags, "-o", objectfile, sourcefile))
+function cache_compargv(self, sourcefile, objectfile, flags, cache_tool)
+    return cache_tool.cmdargv(self:program(), table.join("-c", flags, "-o", objectfile, sourcefile))
 end
 
 -- compile the source file
 function compile(self, sourcefile, objectfile, dependinfo, flags, opt)
+    opt = opt or {}
 
     -- ensure the object directory
     os.mkdir(path.directory(objectfile))
@@ -301,12 +302,14 @@ function compile(self, sourcefile, objectfile, dependinfo, flags, opt)
     try
     {
         function ()
+            local cache_tool = compiler_cache.get(opt.target)
+
             -- generate includes file
             local compflags = flags
             if depfile then
                 -- ccache report "preprocessor error" with "-MMD -MF $depfile" / "-MD -MF $depfile" args
                 -- @see https://github.com/TOMO-CAT/xmake/issues/120
-                if not ccache.is_enabled() and _has_flags_mmd_mf(self) then
+                if not cache_tool and _has_flags_mmd_mf(self) then
                     compflags = table.join(compflags, "-MMD", "-MF", depfile)
                 elseif _has_flags_mm(self) then
                     -- since -MMD is not supported, run nvcc twice
@@ -316,7 +319,7 @@ function compile(self, sourcefile, objectfile, dependinfo, flags, opt)
                     -- replace the wrong *.o file in depfile
                     local wrong_objectfile = path.basename(sourcefile) .. ".o"
                     io.replace(depfile, wrong_objectfile, objectfile, {plain = true})
-                elseif not ccache.is_enabled() and _has_flags_md_mf(self) then
+                elseif not cache_tool and _has_flags_md_mf(self) then
                     -- on windows only -MD and -M are supported
                     compflags = table.join(compflags, "-MD", "-MF", depfile)
                 elseif _has_flags_m(self) then
@@ -331,12 +334,13 @@ function compile(self, sourcefile, objectfile, dependinfo, flags, opt)
             end
 
             -- do compile
-            if ccache.is_enabled() then
-                local program, argv = ccache_compargv(self, sourcefile, objectfile, compflags)
+            if cache_tool then
+                local program, argv = cache_compargv(self, sourcefile, objectfile, compflags, cache_tool)
                 local compile_start_time = os.mclock()
                 local outdata, errdata = os.iorunv(program, argv, {envs = self:runenvs()})
-                local compile_time = os.mclock() - compile_start_time
-                ccache.report_metrics(sourcefile, compile_time)
+                if cache_tool.report_metrics then
+                    cache_tool.report_metrics(sourcefile, os.mclock() - compile_start_time)
+                end
                 return (outdata or "") .. (errdata or "")
             else
                 local program, argv = compargv(self, sourcefile, objectfile, compflags)