+++ xmake/plugins/config-update/xmake.lua
@@ -198,6 +198,7 @@ set_menu {
     { category = "Other Configuration" },
     { nil,                             "debugger",  "kv", "auto",     "Set debugger" },
     { nil,                             "ccache",    "kv", true,       "Enable or disable the c/c++ compiler ccache." },
+    { nil,                             "sccache",   "kv", true,       "Enable or disable the c/c++ compiler sccache." },
     { nil,                             "xcache",    "kv", true,       "Enable or disable the c/c++ compiler xmake cache." },
     { nil,                             "debug",     "kv", false,      "Enable or disable to print debug information." },
     { nil,                             "xcachedir", "kv", nil,        "Set the xcache directory." },